requirements.txt
tensorflow
mlflow
python-box
ensure
pandas 
dvc
notebook
numpy
matplotlib
seaborn
pyYAML
tqdm
joblib
types-PyYAML
scipy
Flask
Flask-Cors
gdown
-e .